Extract JobCard into a generic Card component
This is done because other pages such as the machines page will use the same style of card. So we extract JobCard to reduce code duplication.
